Update placeholder domain for 2023
Placehold.it no longer resolves to a service which generates placeholder
images in 2023, instead redirecting to a private real estate company.
This updates the URL generated for placeholder images to place-hold.it

As of this commit, this doesn't have the approval from the owner of
place-hold.it, and so is only for private use.